#275 workaround to prevent double start of the app. Later it should be nicely structured.
